Isolate loginUser controller tests
